--- sdk/communication/communication-call-automation/test/public/callAutomationClient.spec.ts
@@ -1,113 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import { Recorder } from "@azure-tools/test-recorder";
-import { CommunicationUserIdentifier } from "@azure/communication-common";
-import { assert } from "chai";
-import { Context } from "mocha";
-import { CallAutomationClient, CallInvite, CallConnection } from "../../src";
-import {
-  createRecorder,
-  createTestUser,
-  dispatcherCallback,
-  serviceBusWithNewCall,
-  createCallAutomationClient,
-  waitForIncomingCallContext,
-  waitForEvent,
-} from "./utils/recordedClient";
-import {
-  events,
-  serviceBusReceivers,
-  incomingCallContexts,
-  loadPersistedEvents,
-  persistEvents,
-} from "./utils/recordedClient";
-
-describe("Call Automation Main Client Live Tests", function () {
-  let recorder: Recorder;
-  let callAutomationClient: CallAutomationClient;
-  let callConnection: CallConnection;
-  let testUser: CommunicationUserIdentifier;
-  let testUser2: CommunicationUserIdentifier;
-  let testName: string;
-
-  beforeEach(async function (this: Context) {
-    recorder = await createRecorder(this.currentTest);
-    testUser = await createTestUser(recorder);
-    testUser2 = await createTestUser(recorder);
-    callAutomationClient = createCallAutomationClient(recorder, testUser);
-  });
-
-  afterEach(async function (this: Context) {
-    persistEvents(testName);
-    if (callConnection) {
-      try {
-        await callConnection.hangUp(true);
-      } catch (e) {
-        console.log("Call is terminated");
-      }
-    }
-    serviceBusReceivers.forEach((receiver) => {
-      receiver.close();
-    });
-    events.forEach((callConnectionEvents) => {
-      callConnectionEvents.clear();
-    });
-    events.clear();
-    serviceBusReceivers.clear();
-    incomingCallContexts.clear();
-    await recorder.stop();
-  });
-
-  it("Create a call and hangup", async function () {
-    testName = this.test?.fullTitle()
-      ? this.test?.fullTitle().replace(/ /g, "_")
-      : "create_call_and_hang_up";
-    await loadPersistedEvents(testName);
-
-    const callInvite = new CallInvite(testUser2);
-    const uniqueId = await serviceBusWithNewCall(testUser, testUser2);
-    const callBackUrl: string = dispatcherCallback + `?q=${uniqueId}`;
-
-    const result = await callAutomationClient.createCall(callInvite, callBackUrl);
-    const incomingCallContext = await waitForIncomingCallContext(uniqueId, 8000);
-    const callConnectionId: string = result.callConnectionProperties.callConnectionId
-      ? result.callConnectionProperties.callConnectionId
-      : "";
-    assert.isDefined(incomingCallContext);
-
-    if (incomingCallContext) {
-      await callAutomationClient.answerCall(incomingCallContext, callBackUrl);
-    }
-    const callConnectedEvent = await waitForEvent("CallConnected", callConnectionId, 8000);
-    assert.isDefined(callConnectedEvent);
-    callConnection = result.callConnection;
-
-    await callConnection.hangUp(true);
-    const callDisconnectedEvent = await waitForEvent("CallDisconnected", callConnectionId, 8000);
-    assert.isDefined(callDisconnectedEvent);
-  }).timeout(60000);
-
-  it("Reject call", async function () {
-    testName = this.test?.fullTitle() ? this.test?.fullTitle().replace(/ /g, "_") : "reject_call";
-    await loadPersistedEvents(testName);
-
-    const callInvite = new CallInvite(testUser2);
-    const uniqueId = await serviceBusWithNewCall(testUser, testUser2);
-    const callBackUrl: string = dispatcherCallback + `?q=${uniqueId}`;
-
-    const result = await callAutomationClient.createCall(callInvite, callBackUrl);
-    const incomingCallContext = await waitForIncomingCallContext(uniqueId, 8000);
-    const callConnectionId: string = result.callConnectionProperties.callConnectionId
-      ? result.callConnectionProperties.callConnectionId
-      : "";
-    assert.isDefined(incomingCallContext);
-
-    if (incomingCallContext) {
-      await callAutomationClient.rejectCall(incomingCallContext);
-    }
-
-    const callDisconnectedEvent = await waitForEvent("CallDisconnected", callConnectionId, 8000);
-    assert.isDefined(callDisconnectedEvent);
-  }).timeout(60000);
-});
