+++ sdk/communication/communication-call-automation/test/utils/connectionUtils.ts
@@ -1,6 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { isNode } from "@azure/core-util";
+
+export const baseUri = "https://contoso.api.fake";
+
 export const MOCK_ENDPOINT = "https://REDACTED.communication.azure.com/";
 export const MOCK_CONNECTION_STRING = `endpoint=${MOCK_ENDPOINT};accesskey=eyJhbG`;
 export const CALL_CONNECTION_ID = "callConnectionId";
@@ -14,3 +18,13 @@ export const CALL_CALLBACK_URL = "https://REDACTED.com/events";
 export const CALL_INCOMING_CALL_CONTEXT = "eyJhbGciOiJub25lIiwidHlwIjoiSldUIn0.REDACTED";
 export const CALL_OPERATION_CONTEXT = "operationContext";
 export const MEDIA_SUBSCRIPTION_ID = "mediaSubscriptionId";
+
+declare function btoa(stringToEncode: string): string;
+
+export const generateToken = (): string => {
+  const validForMinutes = 60;
+  const expiresOn = (Date.now() + validForMinutes * 60 * 1000) / 1000;
+  const tokenString = JSON.stringify({ exp: expiresOn });
+  const base64Token = isNode ? Buffer.from(tokenString).toString("base64") : btoa(tokenString);
+  return `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.${base64Token}.adM-ddBZZlQ1WlN3pdPBOF5G4Wh9iZpxNP_fSvpF4cWs`;
+};
